Use custom hook to get active settings
