+++ src/components/ResultOrderer/ResultOrderer.js
@@ -5,7 +5,7 @@ import { FormattedMessage } from 'react-intl';
 import {
   FormControl, Select, Typography,
 } from '@material-ui/core';
-import { useSelector } from 'react-redux';
+import useAcccessibilitySettings from '../../utils/accessibilitySettingsActive';
 
 const allowedDirections = [
   'asc',
@@ -38,13 +38,7 @@ const ResultOrderer = ({
   setOrder,
   userLocation,
 }) => {
-  const userSettings = useSelector(state => state.settings);
-  const accessibiliySettingsLength = [
-    userSettings.mobility,
-    userSettings.colorblind,
-    userSettings.hearingAid,
-    userSettings.visuallyImpaired,
-  ].filter(i => (i !== false && i !== null)).length;
+  const accessibiliySettingsLength = useAcccessibilitySettings().length;
 
   const isValidDirection = direction => direction && allowedDirections.indexOf(direction) > -1;
